I made some changes...
Emulator is the main object (core)  that holds: cpu, gpu, memory.... 

The FileManager is just to load (and raise info for debugging headers) headers and define the type of file we load.

Cames back to the emulator (core), now we knew the type of media so we can initiate the: cpu, gpu, memory...

And finally runs it!

At the moment just the elf32 is implemented, the pbp is a format?

So when I need run I just "run" the emulator... or pause or stop...

ps: don't forget the Settings.xml